requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
Faker==15.1.1
Markdown==3.4.1
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
wrapt==1.14.1
